refactor(saml): move events methods in their own files
