+++ plugins/auth-saml/src/events/create-server-docker-compose.ts
@@ -1,0 +1,13 @@
+import {
+  DsgContext,
+  CreateServerDockerComposeParams,
+} from "@amplication/code-gen-types";
+import { updateDockerComposeProperties } from "../constants";
+
+export function beforeCreateDockerComposeFile(
+  dsgContext: DsgContext,
+  eventParams: CreateServerDockerComposeParams,
+): CreateServerDockerComposeParams {
+  eventParams.updateProperties.push(...updateDockerComposeProperties);
+  return eventParams;
+}
