+++ plugins/auth-saml/src/events/index.ts
@@ -1,2 +1,8 @@
 export * from "./create-admin-app";
 export * from "./create-adminui";
+export * from "./create-server-auth";
+export * from "./create-server-dot-env";
+export * from "./create-server-package-json";
+export * from "./create-server";
+export * from "./create-server-docker-compose";
+export * from "./create-server-secrets-manager";
